Added PayPal API Integration The paypal api was added to the project, in order to keep secrets safe this commit DOES NOT include clientID + secretID which are files needed to perform this operation. However they are not stricly neccesary to run the application or at least they werent when i tested them.
The way this works is that a new paypal activity is called when the paypal button is preseed and confirmed in the payment menu, then latter our token is loaded and the user has to click the text to continue. (I will fix this in the next commit) however just a test this works and works fine. Then a webbased popup appears and the user is prompet to log in or cancel the payment, if the user agrees then the payment is send to the business and the order details are saved in our SQLite database.

Is also important to mention that the post request takes data from the management cart, so the payment matches the order. I am still thinking if i should add more data retrieved from the order (user email, user name, date, etc).